[level 2] Title: [3차] 파일명 정렬, Time: 161.56 ms, Memory: 116 MB -BaekjoonHub

+++ 프로그래머스/2/17686. ［3차］ 파일명 정렬/［3차］ 파일명 정렬.java
@@ -5,7 +5,7 @@ class Solution {
         Arrays.sort(files, Comparator.comparing(
             s -> ((String) s).toLowerCase().replaceAll("\\d.*",""))
                 .thenComparing(
-                    s -> Long.parseLong(((String) s).replaceAll(".*?(\\d+).*", "$1"))));
+                    s -> Integer.parseInt(((String) s).replaceAll(".*?(\\d+).*", "$1"))));
         
         return files;
     }
